backend/requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.0
python-dotenv>=1.0.1
openai>=1.30.0
anthropic>=0.21.0
google-generativeai>=0.7.0
pillow>=10.0.0
python-multipart>=0.0.9
pydantic>=2.0.0
core_logging